# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Services Project",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate with email and password",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "user email",
                                        "default": ""
                                    },
                                    "password": {
                                        "description": "user password",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "invalid credentials"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create new superuser",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "user email",
                                        "default": ""
                                    },
                                    "password": {
                                        "description": "user password",
                                        "default": ""
                                    },
                                    "name": {
                                        "description": "user name",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "401": {
                        "description": "invalid credentials"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "should send the Authorization header with \"Bearer <jwt_token>\"",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}